worker/requirements.txt
# Core
asyncpg>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0

# MLX (Apple Silicon local inference)
mlx>=0.21.0
mlx-lm>=0.21.0
mlx-vlm>=0.1.0

# Rendering (compositor)
playwright>=1.48.0

# Image processing
Pillow>=10.0.0
rembg>=2.0.0
